requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
Werkzeug==3.1.3
